streaming.py: add option to pass metadata to returned messages
Add metadata option to AlertBroker, which when set, allows
protocol-specific metadata alongside message contents to
be returned from the stream.

In order to avoid complications, had to remove the message index
being returned, filter now expects two arguments; the message
and the metadata regardless of whether it will be returned.

Additionally, the handling of heartbeats happens further upstream
to simplify the return type of messages from the stream.