Raise a ValueError instead of ArgumentError for bad Kafka URL
This library attempts to be middleware which can be used in any
context - either in a CLI, or in a headless process. ArgumentError is
not quite right here - and we can't construct one anyway, since we
don't have access to an 'argument.' As a result, the code prior to
this commit would actually raise a NameError because 'self' is not
defined.

Fixes #11.