Add docker-based integration test
Add a test, which is run with 'make test', which starts a Docker
container running the SCiMMA stack's Kafka-in-a-box. Use that Docker
container to validate that the AlertBroker produces and consumes
messages correctly.

At this time, it looks like it produces correctly, but consumption
might not work.

Documentation is still TODO.