Use certifi's root certificates for SSL by default
Kafka brokers on the internet will use SSL typically. Connecting to
them requires verifying the server's SSL certificate. That
verification will only succeed if OpenSSL thinks the certificate looks
legit, which will only succeed if the cert is signed in a chain of
trust that ultimately can be traced back to a known-good root
certificate authority.

On some platforms (notably CentOS 7), the default certificate chain
does not verify Lets Encrypt certificates. This leads to confusing
behavior for clients.

Updating root certificates is usually an OS-specific maneuver, but
certifi is a Python package which distributes Mozilla's set of
known-good root certificates. This package is up-to-date and
well-maintained, so using it should improve our package's portability.
By default, we can use certifi's bundle of certs, and then most
brokers with legitimate SSL server certificates will be verified
correctly.